изменила некоторые поля

+++ src/main/java/ru/work/graduatework/dto/ImageDto.java
@@ -6,13 +6,13 @@ import ru.work.graduatework.Entity.Ads;
 @Data
 public class ImageDto {
 
-    private int id;
+    private long id;
     private String filePath;
     private Long fileSize;
     private String mediaType;
     private byte[] data;
 
-    public ImageDto(int id, String filePath, Long fileSize, String mediaType, byte[] data) {
+    public ImageDto(long id, String filePath, Long fileSize, String mediaType, byte[] data) {
         this.id = id;
         this.filePath = filePath;
         this.fileSize = fileSize;
